Remove "Learn more" button since it doesn't do anything yet
There is no reason to have it if it doesn't do anything, and I don't
even think a "Learn more" button is needed here. The whole pages purpose
is to learn more about Tox.

One option would be a "scroll down" button which is an arrow like (v) or
something, but that's another task.