Converted to use Pelican static site generator
It's rather inconvenient that when you want to edit the footer, you
have to edit it in all of the HTML files as each of them contains a copy
of it. It's easy to make a mistake doing so, which was already proved
before[1], and, frankly, it's quite tedious. Using a template language
for our website would solve this as it, along with many other helpful
features, allows for one instance of HTML to be automatically inserted
into many files.

Since our website is static, I have looked into available static website
generators. The following search criteria was used to evaluate static
site generators

- Being able to generate general purpose websites, not just blogs
- Have good documentation and examples
- Have a reasonably sized community
- Support i18n
- Have small installation footprint (including dependencies)
- Be in Debian package repository
- Use a language that is quite common and easy to pick up

Pelican[2] was chosen as the static website generator as it met all of
the criteria. In fact, even the Jinja2[3], which is the library for
Jinja2 template language used by Pelican, meets all of the criteria, but
the Pelican was selected over bare Jinja2 as using Jinja2 library would
require some boilerplate code to be written, which is already done in
Pelican, and Pelican has quite a few helpful plugins, for example
i18n_subsites[4] and assets[5], which in themselves are boilerplate
codes for getting Jinja2 i18n[6] and assets[7] extensions to work that
we otherwise would have to write ourselves. It's also possible to use
Pelican for the blog, which would allow the website and the blog share
the same templates / web design, if we chose to get away from using
WordPress at some point. However, since we don't use any blogging
functionality Pelican provides, we can think of it as a boilerplate on
top of Jinja2 for now.

This PR converts our website to use Jinja2 templates and Pelican static
website generator. There are a few changes that were done to the website.

- Title of main page says "A New Kind of Instant Messaging" and titles
  of supplementary pages say "<page name> - Tox", e.g. "About - Tox",
  when all pages said "A New Kind of Instant Messaging" before
- Assets (css, js, images, fonts) are moved into `themes` subdirectory
  as this is what Pelican does
- Removed symbolic links to pages

[1] 2756255fa47155f2d2f750e587b9c24a5c9da9c7
[2] http://docs.getpelican.com/en/latest/index.html
[3] http://jinja.pocoo.org/
[4] https://github.com/getpelican/pelican-plugins/tree/master/
    i18n_subsites
[5] https://github.com/getpelican/pelican-plugins/tree/master/assets
[6] http://jinja.pocoo.org/docs/dev/extensions/#i18n-extension
[7] http://webassets.readthedocs.io/en/latest/integration/jinja2.html