Replace Read More Link plugin with template code
The plugin is changing summary's HTML in an unexpected way, breaking
internal content linking functionality. Article HTML is not a valid
HTML, we have some URL starting with '{filename}', whcih Pelican is
supposed substitute with the right path somewhere down the processing
pipeline, but the plugin is run before that substitution takes place,
it uses libxml to modify the HTML, libxml freaks out about '{filename}'
as '{' and '}' are not allowed characters, and ends up escaping them, or
so it seems, which in turn means that Pelican wouldn't do the
substitution and we will end up with '{filename}' URLs in the output.

This plugin has caused quite some issues in the past, it brings in
libxml dependency and making it undo libxml's escaping of exactly
'{filename}' feels like a dirty hack, because next time we might want
to use '{foo}' and will face this same problem again, so I don't think
this plugin is worth keeping. Not to mention that we don't really need
this plugin in the first place, we can make a Read More link without it.
Also, this plugin makes the Read More link be a part of the summary
text, which is undesired since the summary is used to generate the feeds
and our Read More would seem very out of place in there, the feeds
likely put their own Read More in there, so it would be a double Read
More, and reading more twice is not reading more anymore, it's reading
much, too much in fact, assuming you read the whole article twice.